fix warnings

+++ multi_bot/Bot/NetAvatar.h
@@ -8,7 +8,7 @@ struct NetAvatar {
     std::string Name, CountryFlag;
     float PosX, PosY;
     float LastPosX, LastPosY;
-    float WidthSize, HeightSize;
+    int32_t WidthSize, HeightSize;
     int32_t NetID;
     uint32_t UserID;
     bool IsInvis, FacingLeft;
